Done styling Tank

+++ app/Http/Controllers/TankController.php
@@ -2,11 +2,15 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
+use App\Models\Product;
+use App\Models\Tank;
+
 use Carbon\Carbon;
 
-use App\Models\Tank;
-use App\Models\Product;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Hash;
+
 
 use Debugbar;
 
@@ -19,7 +23,8 @@ class TankController extends Controller
         if ($request->onlyName){
             return Tank::select('id','name')->get();
         }
-        else return Tank::all()->map(function($tank){
+
+        $tanks = Tank::all()->map(function($tank){
             return [
                 'id' => $tank->id,
                 'name' => $tank->name,
@@ -28,6 +33,13 @@ class TankController extends Controller
                 'productId' => $tank->product->id,
             ];
         });
+
+        $products = Product::select('id','name')->get();
+
+        return [
+            'tanks' => $tanks,
+            'products' => $products,
+        ];
     }
 
     public function addTank(Request $request){
@@ -41,19 +53,19 @@ class TankController extends Controller
         //Query parameters
         $request->validate([
             'name' => ['required'],
-            'productId' => ['required','exists:products,id'],
+            'product' => ['required','exists:products,id'],
             'stock' => ['required','integer','min:0'],
         ], $messages);
 
         //todo: test if product unavailable
-        $product = Product::find($request->productId);
+        $product = Product::findOrFail($request->product);
 
         $history = collect(
             [
                 [
                     'timestamp' => Carbon::today(),
                     'name' => $request->name,
-                    'productId' => $request->productId,
+                    'productId' => $product->id,
                 ]
             ]
         );
@@ -71,7 +83,7 @@ class TankController extends Controller
 
         $attribute = [
             'id' => 'tangki',
-            'productId' => 'produk',
+            'product' => 'produk',
         ];
 
         $messages = [
@@ -85,20 +97,20 @@ class TankController extends Controller
         $request->validate([
             'id' => ['required', 'exists:tanks,id'],
             'name' => ['required'],
-            'productId' => ['required','exists:products,id'],
+            'product' => ['required','exists:products,id'],
             'stock' => ['required','integer','min:0'],
         ], $messages, $attribute);
 
         //todo: test if product unavailable
-        $product = Product::find($request->productId);
+        $product = Product::find($request->product);
         $tank = Tank::findOrFail($request->id);
 
-        if (($tank->name !== $request->name) || ($tank->product_id !== $request->productId)){
+        if (($tank->name !== $request->name) || ($tank->product_id !== $request->product)){
             Debugbar::info('terpanggil');
             $newHistory = $tank->history->push([
                 'timestamp' => Carbon::today(),
                 'name' => $request->name,
-                'productId' => $request->productId,
+                'productId' => $product->id,
             ]);
         }
 
@@ -109,18 +121,23 @@ class TankController extends Controller
 
         Debugbar::info($newHistory);
         $tank->history = $newHistory;
-        $tank->save();
         $tank->product()->associate($product);
+        $tank->save();
         return ['name' => $product->name];
     }
 
     public function delete(Request $request){
+        $deleter = Auth::user();
         $tank = Tank::findOrFail($request->id);
-        abort_unless($tank->nozzles->isEmpty(), 403, 'Anda harus menghapus atau mengalihkan semua nozzle yang berkaitan dengan tangki ini sebelum menghapus');
-        $tank->delete();
-        return [
-            'name' => $tank->name,
-            'message' => 'Tangki berhasil dihapus',
-        ];
+        if (Hash::check($request->password, $deleter->password)){
+            abort_unless($tank->nozzles->isEmpty(), 403, 'Anda harus menghapus atau mengalihkan semua nozzle yang berkaitan dengan tangki ini sebelum menghapus');
+            $tank->delete();
+            return [
+                'name' => $tank->name,
+                'message' => 'Tangki berhasil dihapus',
+            ];
+        } else {
+            abort(422, 'Password salah');
+        }
     }
 }
